fixed signed unsigned warning in WorkListHelpers push () method of all WorkList types now takes const reference as a parameter instead of copy
git-svn-id: https://svn.ices.utexas.edu/repos/iss/projects/GaloisCpp/trunk@9841 522c16a8-e32f-0410-b4cc-8071c0b41197

+++ include/Galois/Runtime/WorkListHelpers.h
@@ -47,8 +47,8 @@ class FixedSizeRing :private boost::noncopyable, private LL::PaddedLock<concurre
 
   char datac[sizeof(T[__chunksize + 1])] __attribute__ ((aligned (__alignof__(T))));
 
-  T* at(int i) {
-    assert((unsigned)i < (__chunksize + 1));
+  T* at(unsigned i) {
+    assert(i < (__chunksize + 1));
     T* s = reinterpret_cast<T*>(&datac[0]);
     return &s[i];
   }
@@ -304,3 +304,4 @@ struct DummyIndexer: public std::unary_function<const T&,unsigned> {
 
 
 #endif
+
